Remove workflow manager
